Tugas 1/requirement.txt
PyOpenGL==3.1.1
